pyproject.toml
[tool.poetry]
name = "test-app"
version = "0.1.0"
description = ""
authors = ["henri-edh <henri@edhsport.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mistralai = "^0.1.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
